+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Video: Fix background color for fullscreen video playback (#198)
 * Internal: Refactor Modal docs to kill StateRecorder (#199)
 * Internal: Add eslint-plugin-eslint-comments with recommended settings (#200)
+* Video: Makes `aspectRatio` a required prop for `Video` (#201)
 
 ### Patch
 * Internal: Add code coverage to PRs (#185)